refactor: extract GameGenreTags component to new file
